refactored
added lat and lng input to the map

added recentering and rendering

imported the searchbar component
