+++ package.json
@@ -23,6 +23,7 @@
     "react-datepicker": "^4.21.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
+    "react-places-autocomplete": "^7.3.0",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
